Fixed compilation issue when clang on windows. It was cause by a warning due to a implicit cast to a double from a unsigned long long.
#ROBOMERGE-AUTHOR: julien.stjean
#ROBOMERGE-SOURCE: CL 18110211 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v889-18060218)

[CL 18110417 by julien stjean in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/ImageWrapper/Private/Formats/TiffImageWrapper.cpp
@@ -42,7 +42,7 @@ namespace UE::ImageWrapper::Private
 			else
 			{
 				// Naive linear interpolation
-				return TNumericLimits<DataTypeDest>::Max() * (double(ReadedValue) / TNumericLimits<DataTypeSrc>::Max());
+				return TNumericLimits<DataTypeDest>::Max() * (double(ReadedValue) / double(TNumericLimits<DataTypeSrc>::Max()));
 			}
 		}
